+++ src/jtl/Connector/Oxid/Controller/CustomerOrderItem.php
@@ -23,11 +23,44 @@ class CustomerOrderItem extends BaseController {
         $delivery = new CustomerOrderItemModel();
         $delivery->setType('shipping');
         $delivery->setName($data['delName']);
-        $delivery->setPrice($data['OXDELCOST'] - ($data['OXDELCOST'] / 100 * $data['OXDELVAT']));
+        $delivery->setPrice(round($data['OXDELCOST'] / (($data['OXDELVAT'] / 100) + 1 ), 2));
         $delivery->setQuantity(1);
         $delivery->setVat(floatval($data['OXDELVAT']));
 
         $return[] = $delivery;
+
+        if (floatval($data['OXPAYCOST']) > 0) {
+            $payment = new CustomerOrderItemModel();
+            $payment->setType('product');
+            $payment->setName($data['payType']);
+            $payment->setPrice(round($data['OXPAYCOST'] / (($data['OXPAYVAT'] / 100) + 1 ), 2));
+            $payment->setQuantity(1);
+            $payment->setVat(floatval($data['OXPAYVAT']));
+
+            $return[] = $payment;
+        }
+
+        if (floatval($data['OXVOUCHERDISCOUNT']) > 0) {
+            $discount = new CustomerOrderItemModel();
+            $discount->setType('product');
+            $discount->setName('Gutschein Rabatt');
+            $discount->setPrice(round($data['OXVOUCHERDISCOUNT'] * -1, 2));
+            $discount->setQuantity(1);
+            $discount->setVat(0);
+
+            $return[] = $discount;
+        }
+
+        if (floatval($data['OXGIFTCARDCOST']) > 0) {
+            $card = new CustomerOrderItemModel();
+            $card->setType('product');
+            $card->setName('Grußkarte: '.$data['OXCARDTEXT']);
+            $card->setPrice(round($data['OXGIFTCARDCOST'] / (($data['OXGIFTCARDVAT'] / 100) + 1 ), 2));
+            $card->setQuantity(1);
+            $card->setVat(floatval($data['OXGIFTCARDVAT']));
+
+            $return[] = $card;
+        }
 
         return $return;
     }